backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-environ==0.11.2
mysqlclient==2.2.4
PyJWT==2.8.0
celery==5.3.1
redis==4.6.0
Pillow==10.4.0
django-storages==1.14.2
boto3==1.34.144
gunicorn==22.0.0
whitenoise==6.5.0
requests==2.32.3
python-decouple==3.8
django-filter==24.2
django-extensions==3.2.3
djoser==2.2.2
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
django-allauth==0.63.6
django-crispy-forms==2.2
crispy-bootstrap4==2024.1
django-cleanup==8.1.0
django-imagekit==5.0.0
django-mptt==0.16.0
django-taggit==5.0.1
django-activity-stream==2.0.0
django-notifications-hq==1.8.3
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
uvicorn==0.30.1
fastapi==0.111.0
pydantic==2.8.2
sqlalchemy==2.0.31
alembic==1.13.2
pytest==8.2.2
pytest-django==4.8.0
factory-boy==3.3.0
fake2db==2.4.1
coverage==7.6.0
black==24.4.2
flake8==7.1.0
mypy==1.10.1
pre-commit==3.7.1
geopy==2.4.1
# Additional packages for AWS deployment
sentry-sdk==2.17.0
django-health-check==3.18.3
psycopg2-binary==2.9.9
django-ratelimit==4.1.0
